+++ Backend/schoolManagement/src/main/java/com/app/entities/Timetable.java
@@ -24,7 +24,7 @@ import lombok.ToString;
 public class Timetable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int tt_id;
+	private Long tt_id;
 
 	@OneToOne
 	@JoinColumn(name = "classroom_id", referencedColumnName = "classroom_id")
